refactor(connect-ui): adopt vitest-browser-react for the a11y suite
Move the Connect UI accessibility suite onto Browser Mode-native tooling
and tidy the test helpers:

- Render via vitest-browser-react (auto-unmounts between tests); user
  events via vitest/browser; queries via @testing-library/dom. Drop
  @testing-library/react and @testing-library/user-event.
- Split the single test-helper module into dedicated files: render.tsx
  (renderApp), a11y.ts (axe assertion), fixtures.ts (test data).
- Reset the store inside renderApp instead of an afterEach cleanup.
- Relax @typescript-eslint/consistent-type-imports for test files so the
  vi.mock importActual<typeof import(...)> pattern is allowed.
