+++ packages/connect-ui/package.json
@@ -27,8 +27,6 @@
         "@tanstack/react-router": "1.98.3",
         "@testing-library/dom": "10.4.1",
         "@testing-library/jest-dom": "6.9.1",
-        "@testing-library/react": "16.3.2",
-        "@testing-library/user-event": "14.6.1",
         "@types/react": "18.3.3",
         "@types/react-dom": "18.3.0",
         "@vitejs/plugin-react-swc": "4.3.1",
@@ -54,6 +52,7 @@
         "vite": "8.0.16",
         "vite-plugin-svgr": "5.2.0",
         "vitest": "4.1.8",
+        "vitest-browser-react": "2.2.0",
         "zod": "4.0.5",
         "zustand": "5.0.3"
     }
