+++ packages/connect-ui/src/components/LoadingView.test.tsx
@@ -1,12 +1,13 @@
-import { render, screen } from '@testing-library/react';
+import { screen } from '@testing-library/dom';
 import { describe, expect, it } from 'vitest';
+import { render } from 'vitest-browser-react';
 
 import { LoadingView } from '@/components/LoadingView';
 import { I18nProvider } from '@/lib/i18n';
-import { expectAccessibleInBothThemes } from '@/test/harness';
+import { expectAccessibleInBothThemes } from '@/test/a11y';
 
-function renderLoadingView(): HTMLElement {
-    const { container } = render(
+async function renderLoadingView(): Promise<HTMLElement> {
+    const { container } = await render(
         <I18nProvider defaultLanguage="en">
             <LoadingView />
         </I18nProvider>
@@ -14,18 +15,14 @@ function renderLoadingView(): HTMLElement {
     return container;
 }
 
-// LoadingView is shown during suspense/loading — a surface the four main views never expose.
 describe('LoadingView', () => {
-    // NAN-5906 #9: the loading container carries an `aria-label` on a role-less <div>, which screen
-    // readers ignore. axe only downgrades this to "incomplete" (it has child content), so we assert
-    // the fix's contract directly: a status live region announces loading.
-    it('exposes a status region so screen readers announce loading', () => {
-        renderLoadingView();
+    it('exposes a status region so screen readers announce loading', async () => {
+        await renderLoadingView();
 
         expect(screen.getByRole('status')).toBeInTheDocument();
     });
 
     it('has no other accessibility violations in light or dark mode', async () => {
-        await expectAccessibleInBothThemes(renderLoadingView());
+        await expectAccessibleInBothThemes(await renderLoadingView());
     });
 });
