+++ packages/connect-ui/src/test/setup.ts
@@ -1,30 +1,12 @@
 import '@testing-library/jest-dom/vitest';
-import { cleanup } from '@testing-library/react';
 import { afterEach } from 'vitest';
 
 // Real app stylesheet so Tailwind utilities + theme CSS variables resolve in the browser.
 // Without this, axe's color-contrast checks have no computed colors to evaluate.
 import '@/index.css';
 
-import { useGlobal } from '@/lib/store';
-
+// vitest-browser-react auto-unmounts the rendered tree between tests, and renderApp resets the
+// store on each render — so here we only need to undo any theme class a test left behind.
 afterEach(() => {
-    cleanup();
-    // Reset shared singletons so state never leaks between tests.
-    useGlobal.setState({
-        sessionToken: null,
-        provider: null,
-        integration: null,
-        isDirty: false,
-        isSingleIntegration: false,
-        session: null,
-        nango: null,
-        apiURL: 'https://api.nango.dev',
-        isEmbedded: false,
-        isAuthLink: false,
-        detectClosedAuthWindow: false,
-        isPreview: false,
-        showWatermark: false
-    });
     document.documentElement.classList.remove('dark');
 });
